timer saver refactor
